sanity test fixes
